requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.0.0=pyhd8ed1ab_0
addict=2.4.0=py37h89c1867_0
aiohttp=3.8.1=py37h5e8e339_0
aiosignal=1.2.0=pyhd8ed1ab_0
astroid=2.9.0=py37h06a4308_0
async-timeout=4.0.2=pyhd8ed1ab_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd8ed1ab_0
autopep8=1.6.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py_1
blobfile=1.2.8=pypi_0
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
brotlipy=0.7.0=py37h5e8e339_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.3.29=h06a4308_1
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=h18b612c_1001
certifi=2021.10.8=py37h89c1867_2
cffi=1.15.0=py37hd667e15_1
chardet=3.0.4=py37he5f6b98_1008
charset-normalizer=2.0.11=pyhd8ed1ab_0
click=8.0.3=py37h89c1867_1
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.1=py37hf1a17b8_0
cudatoolkit=11.1.1=h6406543_8
cudnn=8.1.0.77=h90431f1_0
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2.30.0=py_0
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
einops=0.4.0=pyhd8ed1ab_0
expat=2.4.4=h9c3ff4c_0
ffmpeg=4.0.2=ha0c5888_2
filelock=3.4.2=pypi_0
fontconfig=2.13.96=ha180cfb_0
fonttools=4.29.1=py37h5e8e339_0
freeglut=3.2.2=h9c3ff4c_1
freetype=2.10.4=h0708190_1
fribidi=1.0.10=h36c2ea0_0
frozenlist=1.3.0=py37h5e8e339_0
future=0.18.2=pypi_0
giflib=5.2.1=h36c2ea0_2
glib=2.69.1=h4ff587b_1
gmp=6.1.2=hf484d3e_1000
gnutls=3.5.19=h2a4e5f8_1
google-auth=2.6.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.42.0=py37hb27c1af_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=hf484d3e_1000
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=4.10.1=py37h89c1867_0
importlib_metadata=4.10.1=hd8ed1ab_0
iniconfig=1.1.1=pyh9f0ad1d_0
isort=5.9.3=pyhd3eb1b0_0
jasper=2.0.33=ha77e612_0
jbig=2.1=h7f98852_2003
jpeg=9e=h7f98852_0
kiwisolver=1.3.2=py37h2527ec5_1
kmeans-pytorch=0.3=pypi_0
lazy-object-proxy=1.6.0=py37h27cfd23_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libdeflate=1.8=h7f98852_0
libffi=3.3=h58526e2_2
libgcc-ng=11.2.0=h1d223b6_12
libgfortran=3.0.0=1
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libglu=9.0.0=he1b5a44_1001
libiconv=1.16=h516909a_0
libimagequant=2.17.0=h7f98852_1
libopencv=3.4.2=hb342d67_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.19.4=h780b84a_0
libstdcxx-ng=11.2.0=he4da1e4_12
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxml2=2.9.12=h03d6c58_0
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=12.0.1=h4bd325d_1
lpips=0.1.4=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.5.1=py37h89c1867_0
matplotlib-base=3.5.1=py37h1058ff1_0
mccabe=0.6.1=py37_1
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
mmcv-full=1.4.0=py37_cuda111_cudnn81_1
more-itertools=8.12.0=pyhd8ed1ab_0
multidict=6.0.2=py37h5e8e339_0
munkres=1.1.4=pyh9f0ad1d_0
mypy=0.812=pyhd8ed1ab_0
mypy_extensions=0.4.3=py37h89c1867_4
ncurses=6.3=h9c3ff4c_0
nettle=3.3=0
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_1
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
oauthlib=3.2.0=pyhd8ed1ab_0
opencv=3.4.2=py37h6fd60c2_1
openh264=1.8.0=hdbcaa40_1000
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1n=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pcre=8.45=h9c3ff4c_0
pillow=9.0.1=py37hc8ad62e_0
pip=22.0.3=pyhd8ed1ab_0
pixman=0.38.0=h516909a_1003
platformdirs=2.4.0=pyhd3eb1b0_0
pluggy=0.13.1=py37h89c1867_4
protobuf=3.19.4=py37hcd2ae1e_0
psutil=5.9.0=py37h5e8e339_0
pthread-stubs=0.4=h36c2ea0_1001
py=1.11.0=pyh6c4a22f_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodomex=3.12.0=pypi_0
pyjwt=2.3.0=pyhd8ed1ab_1
pylint=2.12.2=py37h06a4308_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyqt=5.9.2=py37hcca6a23_4
pysocks=1.7.1=py37h89c1867_4
pytest=6.1.1=py37_0
python=3.7.10=h12debd9_4
python-dateutil=2.8.2=pyhd8ed1ab_0
python-lmdb=1.3.0=py37hcd2ae1e_0
python_abi=3.7=2_cp37m
pytorch=1.8.2=py3.7_cuda11.1_cudnn8.0.5_0
pytorch-msssim=0.2.1=pypi_0
pytz=2021.3=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=6.0=py37h5e8e339_3
qt=5.9.7=h5867ecd_1
readline=8.1=h46c0cb4_0
regex=2022.3.15=py37h7f8727e_0
requests=2.24.0=py_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-image=0.17.2=py37hdf5156a_0
scipy=1.6.2=py37had2a1c9_1
setuptools=60.8.1=py37h89c1867_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.0=h9cd32fc_0
tbb=2021.5.0=h4bd325d_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py37hf1a17b8_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorboardx=2.4=pyhd8ed1ab_0
thop=0.0.31-2005241907=pypi_0
tifffile=2020.10.1=py37hdd07704_2
timm=0.3.2=pyhd8ed1ab_0
tk=8.6.11=h27826a3_1
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.1=py_0
torch-tb-profiler=0.3.1=pypi_0
torchvision=0.9.2=py37_cu111
tornado=6.1=py37h5e8e339_2
tqdm=4.63.0=pyhd8ed1ab_0
typed-ast=1.4.3=py37h5e8e339_1
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
unicodedata2=14.0.0=py37h5e8e339_0
urllib3=1.25.11=py_0
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
wrapt=1.13.3=py37h7f8727e_2
x264=1!152.20180806=h14c3975_0
xmltodict=0.12.0=pypi_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yacs=0.1.6=py_0
yaml=0.2.5=h7f98852_2
yapf=0.31.0=pyhd3eb1b0_0
yarl=1.7.2=py37h5e8e339_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.2=ha95c52a_0
